requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
crawl4ai>=0.3.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
lxml>=4.9.0
